Added connection to local mysql server.

+++ MazeRunner3/src/HighScore/Score.java
@@ -14,7 +14,7 @@ public class Score {
 		return res;
 	}
 	
-	public double getScr()
+	public int getScr()
 	{
 		return score;
 	}
